// src/renderer/src/components/right-sidebar/checks-panel-link-types.ts
export type ChecksPanelLinkKey =
  | 'activeConnectionId'
  | 'activeReview'
  | 'activeWorktreeId'
  | 'branch'
  | 'fallbackGitHubPRNumber'
  | 'fetchHostedReviewForBranch'
  | 'fetchUpstreamStatus'
  | 'isRemoteOperationActive'
  | 'linkedAzureDevOpsPR'
  | 'linkedBitbucketPR'
  | 'linkedGitLabMR'
  | 'linkedGiteaPR'
  | 'linkedPR'
  | 'openModal'
  | 'ownerSettings'
  | 'pr'
  | 'pushBranch'
  | 'repo'
  | 'setGitStatusRefreshNonce'
  | 'setIsPublishingBranch'
  | 'setIsSyncingBranch'
  | 'setRightSidebarOpen'
  | 'setRightSidebarTab'
  | 'syncBranch'
  | 'updateWorktreeMeta'
  | 'handleOpenPR'
  | 'handleUnlinkPullRequest'
  | 'handleLinkAnotherPullRequest'
  | 'pushBeforeCreatePullRequest'
  | 'handlePublishBranch'
  | 'handleSyncBranch'
  | 'handlePullRequestCreated'
;